renamed/moved from util files
