+++ scenario/executor/stepSetState.go
@@ -2,8 +2,14 @@ package scenexec
 
 import (
 	"errors"
+	"fmt"
+	"math/big"
 
 	scenmodel "github.com/multiversx/mx-chain-scenario-go/scenario/model"
+	worldmock "github.com/multiversx/mx-chain-scenario-go/worldmock"
+	"github.com/multiversx/mx-chain-scenario-go/worldmock/esdtconvert"
+
+	vmcommon "github.com/multiversx/mx-chain-vm-common-go"
 )
 
 // ExecuteSetStateStep executes a SetStateStep.
@@ -100,4 +106,120 @@ func (ae *ScenarioExecutor) UpdateAccount(scenAccount *scenmodel.Account) error 
 
 	ae.World.AcctMap.PutAccount(existingAccount)
 	return nil
+}
+
+func convertAccount(testAcct *scenmodel.Account, world *worldmock.MockWorld) (*worldmock.Account, error) {
+	storage := make(map[string][]byte)
+	for _, stkvp := range testAcct.Storage {
+		key := string(stkvp.Key.Value)
+		storage[key] = stkvp.Value.Value
+	}
+
+	err := esdtconvert.WriteScenariosESDTToStorage(testAcct.ESDTData, storage)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(testAcct.Address.Value) != 32 {
+		return nil, errors.New("bad test: account address should be 32 bytes long")
+	}
+
+	account := &worldmock.Account{
+		Address:         testAcct.Address.Value,
+		Nonce:           testAcct.Nonce.Value,
+		Balance:         big.NewInt(0).Set(testAcct.Balance.Value),
+		BalanceDelta:    big.NewInt(0),
+		DeveloperReward: big.NewInt(0).Set(testAcct.DeveloperReward.Value),
+		Username:        testAcct.Username.Value,
+		Storage:         storage,
+		Code:            testAcct.Code.Value,
+		OwnerAddress:    testAcct.Owner.Value,
+		AsyncCallData:   testAcct.AsyncCallData,
+		ShardID:         uint32(testAcct.Shard.Value),
+		IsSmartContract: len(testAcct.Code.Value) > 0,
+		CodeMetadata: (&vmcommon.CodeMetadata{
+			Payable:     true,
+			Upgradeable: true,
+			Readable:    true,
+		}).ToBytes(), // TODO: add explicit fields in scenario JSON
+		MockWorld: world,
+	}
+
+	return account, nil
+}
+
+func validateSetStateAccount(scenAccount *scenmodel.Account, converted *worldmock.Account) error {
+	err := converted.Validate()
+	if err != nil {
+		return fmt.Errorf(
+			`"setState" step validation failed for account "%s": %w`,
+			scenAccount.Address.Original,
+			err)
+	}
+	return nil
+}
+
+func validateNewAddressMocks(testNAMs []*scenmodel.NewAddressMock) error {
+	for _, testNAM := range testNAMs {
+		if !worldmock.IsSmartContractAddress(testNAM.NewAddress.Value) {
+			return fmt.Errorf(
+				`address in "setState" "newAddresses" field should have SC format: %s`,
+				testNAM.NewAddress.Original)
+		}
+	}
+	return nil
+}
+
+func convertNewAddressMocks(testNAMs []*scenmodel.NewAddressMock) []*worldmock.NewAddressMock {
+	var result []*worldmock.NewAddressMock
+	for _, testNAM := range testNAMs {
+		result = append(result, &worldmock.NewAddressMock{
+			CreatorAddress: testNAM.CreatorAddress.Value,
+			CreatorNonce:   testNAM.CreatorNonce.Value,
+			NewAddress:     testNAM.NewAddress.Value,
+		})
+	}
+	return result
+}
+
+func convertBlockInfo(testBlockInfo *scenmodel.BlockInfo, currentInfo *worldmock.BlockInfo) *worldmock.BlockInfo {
+	if testBlockInfo == nil {
+		return currentInfo
+	}
+
+	if currentInfo == nil {
+		currentInfo = &worldmock.BlockInfo{
+			BlockTimestamp: 0,
+			BlockNonce:     0,
+			BlockRound:     0,
+			BlockEpoch:     0,
+			RandomSeed:     nil,
+		}
+	}
+
+	if !testBlockInfo.BlockTimestamp.OriginalEmpty() {
+		currentInfo.BlockTimestamp = testBlockInfo.BlockTimestamp.Value
+
+	}
+
+	if !testBlockInfo.BlockNonce.OriginalEmpty() {
+		currentInfo.BlockNonce = testBlockInfo.BlockNonce.Value
+	}
+
+	if !testBlockInfo.BlockRound.OriginalEmpty() {
+		currentInfo.BlockRound = testBlockInfo.BlockRound.Value
+	}
+
+	if !testBlockInfo.BlockEpoch.OriginalEmpty() {
+		currentInfo.BlockEpoch = uint32(testBlockInfo.BlockEpoch.Value)
+	}
+
+	if testBlockInfo.BlockRandomSeed != nil && !testBlockInfo.BlockRandomSeed.OriginalEmpty() {
+		var randomsSeed [48]byte
+		copy(randomsSeed[:], testBlockInfo.BlockRandomSeed.Value)
+		currentInfo.RandomSeed = &randomsSeed
+
+	}
+
+	return currentInfo
 }
