+++ scenario/exporter/getSCCode.go
@@ -1,0 +1,17 @@
+package exporter
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+)
+
+// GetSCCode retrieves the bytecode of a WASM module from a file
+func GetSCCode(fileName string) []byte {
+	code, err := os.ReadFile(filepath.Clean(fileName))
+	if err != nil {
+		panic(fmt.Sprintf("GetSCCode(): %s", fileName))
+	}
+
+	return code
+}
