--- scenario/io/util.go
@@ -1,12 +1,0 @@
-package scenio
-
-import (
-	fr "github.com/multiversx/mx-chain-scenario-go/scenario/expression/fileresolver"
-)
-
-// NewDefaultFileResolver yields a new DefaultFileResolver instance.
-// Reexported here to avoid having all external packages importing the parser.
-// DefaultFileResolver is in parse for local tests only.
-func NewDefaultFileResolver() *fr.DefaultFileResolver {
-	return fr.NewDefaultFileResolver()
-}
