+++ scenario/model/multiTransferUtil.go
@@ -1,18 +1,12 @@
-package util
+package scenmodel
 
 import (
-	"fmt"
-	"os"
-	"path/filepath"
-
 	"github.com/multiversx/mx-chain-core-go/core"
-	scenmodel "github.com/multiversx/mx-chain-scenario-go/scenario/model"
 	txDataBuilder "github.com/multiversx/mx-chain-vm-common-go/txDataBuilder"
 )
 
 // CreateMultiTransferData builds data for a multiTransferESDT
-func CreateMultiTransferData(to []byte, esdtData []*scenmodel.ESDTTxData, endpointName string, arguments [][]byte) []byte {
-
+func CreateMultiTransferData(to []byte, esdtData []*ESDTTxData, endpointName string, arguments [][]byte) []byte {
 	multiTransferData := make([]byte, 0)
 	multiTransferData = append(multiTransferData, []byte(core.BuiltInFunctionMultiESDTNFTTransfer)...)
 	tdb := txDataBuilder.NewBuilder()
@@ -34,14 +28,4 @@ func CreateMultiTransferData(to []byte, esdtData []*scenmodel.ESDTTxData, endpoi
 	}
 	multiTransferData = append(multiTransferData, tdb.ToBytes()...)
 	return multiTransferData
-}
-
-// GetSCCode retrieves the bytecode of a WASM module from a file
-func GetSCCode(fileName string) []byte {
-	code, err := os.ReadFile(filepath.Clean(fileName))
-	if err != nil {
-		panic(fmt.Sprintf("GetSCCode(): %s", fileName))
-	}
-
-	return code
 }
